Actually load the dot env file before virtualenv creation.
