+++ tests/unit/test_core.py
@@ -3,7 +3,8 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-from pipenv.core import load_dot_env, warn_in_virtualenv
+from pipenv.core import warn_in_virtualenv
+from pipenv.utils.environment import load_dot_env
 from pipenv.utils.shell import temp_environ
